init.sh: wait for the external SD
Right after the MTD conversion, the device is rebooted.
If the installer zip is inside the external SD and this one is not ready
when recovery looks for it, the device is immediately rebooted without
/emmc/cyanogenmod.cfg being deleted. The device will hence be rebooted
repeatedly until recovery succesfully opens the installer zip.

The while loop is replaced with a simpler "busybox sleep" to wait for both
the memories. Three seconds should be more than needed.