init.sh: don't mount /sdcard
The needed device nodes are not created and /sdcard can't be mounted.

Instead of creating the device nods, the mount command is removed.
This means that the flash will never be automatically re-started
after the conversion if the zip is inside the external SD.

This prevents conflicts with the mountpoints. recovery mounts
automatically the external SD card in /sdcard if extendedcommand is
found. Normally it's mounted in /emmc, which means that the path to
the zip is wrong in this case and without the proper mount commands
the flash will always abort and the device repeatedly rebooted.

Also, unmount /emmc, the content of /dev is deleted right before the
execution of /init and recovery will unmount it anyway.